bkfilter refactor
